requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.2
python-multipart==0.0.9
numpy==1.26.4
pandas==2.2.2
opencv-python==4.10.0.84
Pillow==10.3.0
torch>=2.2.0
torchvision>=0.17.0
timm==1.0.7
safetensors==0.4.3
PyYAML==6.0.1
